Update patch set 1
Patch Set 1:

INFO messages should only be messages that are useful and significant to the user.  Notably it can never be correct to have LOG(INFO);  Especially when it's being used simply to log entry into a function, which as you say is the very definition of what TRACE is for.

Patch-set: 1