Update patch set 1
Patch Set 1:

Note that it turns out boost::log is not available on RHEL7 (it was added in 1.54 and we only have 1.53) so we'll need to come up with a different solution.  I implemented a quick and dirty logger for now.

Patch-set: 1
Reviewer: Gerrit User 46 <46@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0