Delay syncing switch state till agent is connected to its opflex peer
When the agent starts up, if opflex peer is not available, the
agent is not able to download EPGs, policies etc. So when the
agent syncs flow-table state on the switch, it just ends up
removing all the flows. Instead, now the agent waits to know
about the first EPG as an indirect indication of being connected
to peer; sync-ing is started a little after this point.

Change-Id: I3126f9f938a02738b9257c4087fb0f5c65c5df3c
Signed-off-by: Amit Bose <bose@noironetworks.com>