Don't populate 'actions' when deleting flows
When deleting a flow, if the action part
refers to a deleted group id, OVS reports OFPBAC_BAD_OUT_GROUP
("invalid group id"). This typically happens when
syncing switch state after a restart.
Since actions are not really required for deleting a
flow, this change avoid setting actions for deletes.

Change-Id: Ie9cf864b45cdb039f3a49670f142a86d1616d18a
Signed-off-by: Amit Bose <bose@noironetworks.com>