Add support for sending commands to the OVS daemon
The purpose of this change is provide a way to send control
commands to the OVS daemon similar to ovs-appctl. These commands
are sent over JSON-RPC to a UNIX domain socket that the daemon
listens to. This change enhances class SwitchConnection to be
able to communicate over this socket using JSON.

Change-Id: I57e12453fbdaf562f979aa351a9e58bd6be84808
Signed-off-by: Amit Bose <bose@noironetworks.com>