Update patch set 2
Patch Set 2:

(8 comments)

Thanks Deepjyoti, this looks very good.  I have some nitpicky items in the review that should be a couple minutes' work.

I cherry-picked this change out to separate it from your other commits.  If you pull this change set with a --rebase it should fix your tree.

When you update, you'll want to do a git commit --amend to add additional changes.  Make sure the Change-Id: line is present in your commit message, so that when you push the commit it will show up as a separate patch set on this change rather than a new change list.

I'd also appreciate another patch to add the missing #ifdef __cplusplus lines to the other C headers.  We'll also need to add unit tests for the c wrapper before it can be viable, but I won't block this patch on that.

Patch-set: 2
Label: Verified=0