Remove CPU hotspot in URIQueue::proc_thread_func()
perf profiles of the agent with 1000 endpoints show a CPU spike
when endpoint stats are reported. The process function of
URIQueue shows up as a hotspot because of the book-keeping
needed by multi-index container every time an item is popped from
the queue. Instead of popping an item at a time, we now use
swap() and destroy to avoid book-keeping overhead.

Change-Id: Id015452ea229447e55ac89c9bc2f82848ed4231a
Signed-off-by: Amit Bose <bose@noironetworks.com>