Update patch set 1
Patch Set 1:

This also means repeated updates to the same URI won't be consolidated as effectively.  It's obviously a win when handling large numbers of unique updates (such as this case) but may be a problem if the same set of objects is being repeatedly updated (such as when modifying children of an object with a listener).  Those would have to be part of separate mutate operations though to be a problem.  Probably still worth it?

Patch-set: 1
Reviewer: Gerrit User 46 <46@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0