Update patch set 3
Patch Set 3:

(3 comments)

How is this going to be useful in the context of libopflex?  Is the intention that the upstream user of the library would use the inmemoryloghandler and chainedloghandler?  Would this make more sense implemented directly in agent-ovs if that's what we want?

Patch-set: 3
Reviewer: Gerrit User 46 <46@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0