Use private VNIDs for External Networks in flow pipeline
External networks don't have a VNID that is used
for encapsulation on the wire, hence they are not provided
to the agent by opflex peers. But we need them for
policy enforcement in the flow pipeline, so we generate
VNIDs for them making sure that they don't overlap with the
legal range of VNIDs.

Change-Id: I6488741048b207fc8393c5bfcb9cf61823c49c06
Signed-off-by: Amit Bose <bose@noironetworks.com>