Avoid addition of duplicate subnet flows
ACI sets the 'address' field of subnet to the gateway
address of the subnet (instead of retaining only the
significant bits). This is doesn't match When the same
subnet is read from ExtraConfigManager. Thus we attempt
to create multiple (identical) flows for a subnet.
This change fixes the same subnet check in
HandleRoutingDomainUpdate to compare the significant bits
only.

Change-Id: Icd50d9a37347199cc23d1fcb265d36bd6a21ed15
Signed-off-by: Amit Bose <bose@noironetworks.com>