Fix handling of contract removal in PolicyManager
If a contract was removed from MODB, then PolicyManager
would lose track of the providers/consumers of that
contract. Subsequently, if the contract was added back,
it would end up with no providers/consumers.
To fix this issue, the index entry for a contract in
PolicyManager is kept around as long as there is
atleast one provider or consumer, or the contract exists
in MODB.

Change-Id: Ia1813636a76f5197c3475258a17c8be5515984c5
Signed-off-by: Amit Bose <bose@noironetworks.com>