Use flow cookies for (nearly) all flows
This avoids most cases where duplicate policy could cause a flow entry
to be deleted incorrectly if the flow entry was added for multiple
objects and then one of those objects is deleted.  If this happens,
the only way to repair the flow table is to restart the agent.  With
the cookie, the extra flow should survive in the table and only the
correct copy be deleted, and the table should converge to the correct
state.

Change-Id: I7cc55dc90b6504aec1bf6b3f34d0ac3f7f9c1df2
Signed-off-by: Rob Adams <readams@readams.net>