Refactor table state management to handle duplicate matches more robustly.
When multiple objects generate a flow with a duplicate match and
priority, and one of those is then subsequently deleted, then the flow
will be removed and the flow table enters an invalid state.  With this
change, we keep track of any conflicting flows so the table state will
correctly reconverge after the configuration error is corrected.

Also fixes security group configuration in endpoint files and some
minor refactoring.

Change-Id: I1ce4204e8d0db58a762aa7ce7bfd0c6ca65427cc
Signed-off-by: Rob Adams <readams@readams.net>