Update patch set 10
Patch Set 10:

I think we'd need a way to characterize the set of flows we expect for a specific set of stats.

One way would be to build an appropriate index in TableState so the stats manager can query for the set of flows for a given classifier.  Then when building the counter for a classifier, it would wait until it's seen an update for all of the flows.

Patch-set: 10
Label: Code-Review=0