Update patch set 9
Patch Set 9:

(7 comments)

I have some nitpicks but mostly looks pretty good to me.  Could we squash this commit down onto the other one so we don't have to deal with 2 dependent changesets?  I'm not sure we're gaining much with keeping them separate.  It looks like Keith's other changeset (the gitignore change) is already included in this one so that changeset can probably be dropped.

Patch-set: 9