Update patch set 14
Patch Set 14:

(6 comments)

It still looks like just a stub to me.  Did you mean to push a different change set?  Most of my comments are just notes that I hope will be useful as you work on your implementation later.

High level item on indentation:  Indent to 4 spaces, with arguments aligned with parens, brace on same line as function body.  In the patch it's both internally inconsistent and inconsistent with the rest of the code.

example:

namespace ovsagent {

class MyClass {
public:
    MyClass(int tooMany, int argumentsToFit, int onOneLine
            int indentsLikeThis) {
        // stuff
    }

};

} // namespace ovsagent

Finally:  We'll need to work on getting the vpplib built for the build infra so that in these builds we can enable VPP and run the unit tests.  This will involve modifying the jenkins-job-builder task for OpFlex to build a tarball package for vpplib that we can depend on.

Patch-set: 14
Label: Verified=0