Update patch set 2
Patch Set 2: -Code-Review

I think given the way you're written this, the code will be a lot simpler and cleaner if you just combine the stats updating for both the contract and security group stats into a single class.

Also you've still got quite a lot of copy/pasted code and the unit test coverage is still not very good.  I know you're basing it on the existing PolicyStatsManager tests but it would be good to get reasonable coverage here.

Patch-set: 2
Label: Code-Review=0