Update patch set 7
Patch Set 7:

(28 comments)

This looks very close now though I do notice a couple of serious issues as well as a number of style nitpicks that are in the inline comments.

On the topic of the git commit message:  Rather than a changelog of the history of the patch, think of this as a comment that represents the current state.  In addition, in gerrit you really want to have a shorter message as the first line.  For example:

Added new stats manager to track security group statistics

* Refactored PolicyStatsManager to separate common functionality
  into base class
* Refactored existing policy stats manager tests to simplify
  and improve test coverage
* Update flow cookie for security groups to use classifier
  rather than security group.

If you want to make comments on the history of the patch or on the changes you made in a particular revision, you can make a comment on the gerrit change.

Another question for you:  Do you think we should rename PolicyStatsManager to ContractStatsManager and AccStatsManager to SecGroupStatsManager given the way you've separated them out into two classes?  We can also rename BaseStatsManager to just StatsManager.

Patch-set: 7