Rule priority incorrect for opflex trunk ports.
We currently create 2 dhcp rules with same priority 200 for untagged and
tagged traffic on the same port. The result is that we end up hitting
the first rule even for the vlan traffic preventing the sub-interface
from acquiring the dhcp ip address.

Without the fix the rules look like this.
 cookie=0x0, duration=3.957s, table=0, n_packets=0, n_bytes=0, priority=200,udp,in_port=1,tp_src=68,tp_dst=67 actions=load:0x6->NXM_NX_REG7[],goto_table:3
 cookie=0x0, duration=4.299s, table=0, n_packets=0, n_bytes=0, priority=200,udp,in_port=1,dl_vlan=4094,tp_src=68,tp_dst=67 actions=load:0x6->NXM_NX_REG7[],write_metadata:0x1/0xff,goto_table:3

With the fix the second rule would look like this.
 cookie=0x0, duration=4.299s, table=0, n_packets=0, n_bytes=0, priority=201,udp,in_port=1,dl_vlan=4094,tp_src=68,tp_dst=67 actions=load:0x6->NXM_NX_REG7[],write_metadata:0x1/0xff,goto_table:3

Verified the fix creating flows using the attribute access-interface-vlan": 4094 in the ep file and made sure with the attribute the rule is created with priority 201 and without it with priority 200.

Updated unit test files.

Change-Id: I90f27d6940b58b075a36156e6a60497a0688cb6f
Signed-off-by: Madhu Challa <challa@gmail.com>