Update patch set 4
Patch Set 4:

I waited to see if there was a need for InterfaceProfile and Node profile to be visible in Genie.So far, I didn't see a need, so removing it. As for the Bounce endpoint, If the agent asks to resolve the Routing Domain at a later time, OpflexP may not have the object, so maybe it is better to have the vnids directly.

Patch-set: 4