Agent replying to arp with dhcp ip on native vlan.
This happen due to the following rule.

 cookie=0x0, duration=6892.708s, table=3, n_packets=0, n_bytes=0, priority=51,arp,reg4=0x1,reg6=0x1,dl_dst=ff:ff:ff:ff:ff:ff,arp_tpa=192.168.0.2,arp_op=1 actions=move:NXM_OF_ETH_SRC[]->NXM_OF_ETH_DST[],load:0xfa163e5239f4->NXM_OF_ETH_SRC[],load:0x2->NXM_OF_ARP_OP[],move:NXM_NX_ARP_SHA[]->NXM_NX_ARP_THA[],load:0xfa163e5239f4->NXM_NX_ARP_SHA[],move:NXM_OF_ARP_SPA[]->NXM_OF_ARP_TPA[],load:0xc0a80002->NXM_OF_ARP_SPA[],IN_PORT

The goal is to not respond to similar arp request on the fabric uplink.

We solve this by adding a higher priority rule that looks like this, note the match on in_port=<tunnel_port> and action=drop.
 cookie=0x0, duration=6892.708s, table=3, n_packets=0, n_bytes=0, priority=52,arp,reg4=0x1,reg6=0x1,in_port=3,dl_dst=ff:ff:ff:ff:ff:ff,arp_tpa=192.168.0.2,arp_op=1 actions=drop

Change-Id: I79857106ba5b45d35618554629e2d07bf09fe6a9
Signed-off-by: Madhu Challa <challa@gmail.com>
(cherry picked from commit f796286a09f029a7bbd13218c59fbb1dfa92a043)