Enhanced genie model to 1. add a session container for all span artifacts 2. add a reference to span/session instances from EPG
Assumption: opflex proxy will add a reference to EPG pointing to
the relevant span/sessions.
When End points are declared, any references to EPGs will be resolved
and any references from the EPG to other artifacts will be resolved.
On removal of an end point, the EPG is updated by opflex proxy to add/remove
span/session references.
Added span manager to manage span artifacts and logic.
Updated Policies.cpp to generate span artifacts for mock server
and verified that span artifacts are indeed created and resolved by the agent.

Change-Id: I42d18fb49ee3e18c6b19f123ef79ce3a5c11c474
Signed-off-by: tarbash <waris12@yahoo.com>