Removed undefined unused extern mf_fields declaration.
The extern declaration for mf_fields does not refer to
any definition in any of the sources. Also the array itself
is never used either in the source where it is externed
or other sources. This is however, causing symbol reference
errors when running the opflex agent. Hence removal of the
extern declaration should solve the error.

Change-Id: Idce65139d2b05b4a47a03fb27dc29765452c7acb
Signed-off-by: Madhu Mohan Nelemane <mmnelemane@suse.com>