SVI in opflex mode
L3out SVI is created as a BD with an implicit EPG.
No policy is resolved on the SVI, so PolicyManager is unaware.
EndpointManager maintains the BD members.
Traffic within the SVI BD is always allowed.
Traffic to/from other EPGs is not allowed.
Send unknown unicast traffic to uplink regardless of inventory mode.

Signed-off-by: Kiran Shastri <shastrinator@gmail.com>
Change-Id: Ic18c90d4487a37e8786d89631463fe24007651c5