Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)

I agree, since this was also integrated for the purposes of demo and many changes were done for svc-pod to work, that got intertwined. The #ifdef is just to avoid creation of the ep<-->svc code creation and also to avoid few make check issues. But the extra stats table is in place and working fine. I will start working on fixing this part as well. If u want me to split all the svc stats related changes, pls let me know. I can see how bad it is to remove them from this.

Patch-set: 2
Label: Verified=0