Infrastructure for writing ovsdb realted unit tests
Unit tests for testing rpc request response required
infrastructure support, which has been added with this checkin.
The user needs to supply on the wire request response pairs
to be validated by the unit tests. The request is hairpinned
in libopflex and a matching response is written back to the handler.

Signed-off-by: Mandeep S Sandhu <waris12@yahoo.com>
Change-Id: Id9ac1315a597a3b7a665f450245492b75285e221