Permit untagged traffic for access-interface-vlan
Openshift has a special use case where they need networking
during bootstrap from initrd. When such a VM is used with OVS
since the EP has access-interface-vlan OVS would push vlan
tags on ingress and pop vlan tags on egress.

This would fail during bootstrap when they have an IP on
eth0 and not on eth0.<access-interface-vlan>

This can be fixed more cleanly by having tc inside the VM
push and pop the vlans rather than OVS doing it. That way
OVS can always assume untagged and once we properly initialize
the VM after bootstrap we can have tc do the push /pop

They do not have iproute2 package with tc enabled on their VMs

So the fix is to allow tag / untagged on egress
and duplicate packets on ingress to send them as both
tagged and untagged only for access-interface-vlan flows

May be in future we could run our own container with tc
and do the push / pop in the vm and revert this change

Signed-off-by: Madhu Challa <challa@gmail.com>
Change-Id: I58c6977b858ce563f6db00b78d192519720db6ab