Prometheus: annotate pod<-->svc metrics with "scope" of service
Today service metrics in prometheus are created with service attribute's "name" and "namespace". There are cases where internal and external service will have same set of attributes but different service mapping IPs, which leads to same metric pointer allocated by prometheus. Due to this, stats will be incorrect and memory corruption can happen when one of the service files get removed/modified.

Discussed with Sumit. Will be checking UUID of service to see if its internal vs external service for now. And annotate prometheus with "scope: cluster-svc or ext-svc". In future, hostagent will be setting an attribute named "scope" with either "cluster" or "ext" or "nodeport".

Signed-off-by: Gautam Venkataramanan <gautam.chennai@gmail.com>
Change-Id: I993fce89e9e9f24fdf4275a9a7677e1e46a61a6c