Fix for cookie reconciliation
If a flow modification has a cookie change as well,
then delete the old flow and add the new flow.
We can also handle it with a mod, but that path requires passing
more parameters between tablestate and flowexecutor.
Earlier cookie changes were not really taking effect,
all test fixes are related to this issue.

Signed-off-by: Kiran Shastri <shastrinator@gmail.com>
Change-Id: Ib0fa5961c0d18dcfdb24d759b9632dbd6741d840