Do not add L34 flows if remotesub missing
This bug was reported by Cisco IT where a secgrp with
no remote subnets basically allows all ingress traffic
due to how we program the rules. Since this is a day 0
issue we decided to fix the behavior while at the same
time providing an option for someone to revert to the
old behavior.

- If remote subnet is missing do not add L3 and L4 flows
- A new function add_l2classifier_entries takes care of
  adding l2 only flows
- Add :
    "behavior": {
        "l34flows-without-subnet" : true
    }
 to opflex config to override this behavior with old, default is false.
- verified make check passes with / without this option since I added
  default 0.0.0.0/0 ::/0 subnet to relevant secgroups.

Signed-off-by: Madhu Challa <challa@gmail.com>
Change-Id: Iffa03301c24bd576034f5f5858ebb8c75fa72f5e